Set environment variables to only expose allocated devices
Set both HIP and OpenCL environment variables for devices
to be exposed. The device ordinals are calculated from the
sorted PCI addresses. Tested on machines with dual AMD
and NVIDIA cards where the ROCm GPUs in the system were
identified as cards #2 and #4.

Note that for pods not requesting any devices this code will
never be called, so unless you set the default exposed devices
all of them will be exposed.