Node labeller: add non-beta labels
The non-beta label's format will depend on the node setup.  If the node
has different type of GPUs, the labels will have a count as the value.
If the node only has one type of GPU, the label with has the attribute
as the value.

The beta labels use both format.

Change-Id: I3292b5069e2e3c03fe0b5a047b710ce0d1d09949