Change SimpleHealthCheck to use `/sys/class/kfd` (#116)
* Change SimpleHealthCheck to use `/sys/class/kfd`

This modification allows the plugin to avoid /dev/kfd and uses sysfs
instead, which is readable in unprivileged containers. We use both the
cpu_count and the gfx_target_version in a properties file in the
sysfs kfd topology, to determine if a GPU is found on a host node.
When cpu_count == 0 and gfx_target_version is nonzero in a
properties file, then we assert that a GPU is found and the
simpleHealthCheck is true.

* Remove TODO item