Optimize allocator logic for partitions
The current implementation of allocator takes time to do the calculation because of the massive number of subsets possible if many resources are requested.
This is hit when there are partitions involved especially.
This PR optimizes it by reducing the number of subsets when partitions of involved. It makes use of the fact that all partitions of the same original GPU will have same link types, numa node etc.
and avoids including them with each other in subset combinations and also adds more unit tests