Device Plugin changes for GPU Partitions
This commit contains changes made on the plugin side to invoke
previously defined helper functions to start the appropriate number of
plugins

Terms to understand before viewing the changes in this commit:

Homogeneous Node:
   If all GPUs in a node are following the same compute and memory
   partition style, the node is considered homogeneous

Heterogeneous Node:
   If the GPUs on a node have different different compute and memory
   partition styles, the node is considered heterogeneous (Put simply,
   if node is not homogeneous)

From the count of gpu's for each partitionType, device plugin
determines if node is homogeneous or heterogeneous.

-> If node is homogeneous, the ListAndWatch function is the same as it
   was before. One plugin is started for resource last name "gpu"
-> If node is heterogeneous, ListAndWatch sends the devices to their
   respective resource type depending on its partitionType. Multiple
   plugins are started for the different resource types present and are
   identified using resourceLastName (cpx_nps1, spx_nps1, etc.)

Example of how the map looks:
   Device counts: map[cpx_nps1:8 spx_nps1:4]

NOTE:

import "strings" was missed in this previous PR:
https://github.com/ROCm/k8s-device-plugin/pull/116
This is causing "undefined: strings" compilation error.
Added that import here as it was interfering with being able to compile
my changes